Added first step of refactoring
