+++ .gitignore
@@ -4,3 +4,4 @@ venv
 /auto_text
 /corrected_text
 /corrected_figures
+/detection_help_dir
